Fix: Create proper library structure to fix build
- Extract code from main.rs to lib.rs
- Fix module imports in tests.rs
- Update binary to correctly use library functions
- Resolve compilation errors with proper module organization
