+++ src/bin/btram.rs
@@ -1,5 +1,5 @@
 use btram::*;
 
 fn main() -> anyhow::Result<()> {
-    run()
+    btram::run()
 }
